Move new tests to RouterFactoryTest
