+++ tests/RouterFactoryTest.php
@@ -1,0 +1,252 @@
+<?php
+
+
+namespace Yiisoft\Router\Tests;
+
+
+use PHPUnit\Framework\TestCase;
+use Psr\Container\ContainerInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Yiisoft\Router\Group;
+use Yiisoft\Router\MatchingResult;
+use Yiisoft\Router\Route;
+use Yiisoft\Router\RouterFactory;
+use Yiisoft\Router\RouterInterface;
+use Yiisoft\Router\Tests\Support\Container;
+
+final class RouterFactoryTest extends TestCase
+{
+    public function testContainerAutoInjectionByGroupFactory(): void
+    {
+        $container = $this->getContainer();
+
+        $apiGroup = Group::create(
+            '/api',
+            [
+                Route::get('/info')->name('api-info'),
+                Group::create(
+                    '/v2',
+                    [
+                        Route::get('/user')->name('api-v2-user/index'),
+                        Route::get('/user/{id}')->name('api-v2-user/view'),
+                        Group::create(
+                            '/news',
+                            [
+                                Route::get('/post')->name('api-v2-news-post/index'),
+                                Route::get('/post/{id}')->name('api-v2-news-post/view'),
+                            ]
+                        ),
+                        Group::create(
+                            '/blog',
+                            [
+                                Route::get('/post')->name('api-v2-blog-post/index'),
+                                Route::get('/post/{id}')->name('api-v2-blog-post/view'),
+                            ]
+                        ),
+                        Route::get('/note')->name('api-v2-note/index'),
+                        Route::get('/note/{id}')->name('api-v2-note/view'),
+                    ]
+                ),
+                Group::create(
+                    '/v2',
+                    [
+                        Route::get('/user')->name('api-v2-user/index'),
+                        Route::get('/user/{id}')->name('api-v2-user/view'),
+                        Group::create(
+                            '/news',
+                            [
+                                Route::get('/post')->name('api-v2-news-post/index'),
+                                Route::get('/post/{id}')->name('api-v2-news-post/view'),
+                            ]
+                        ),
+                        Group::create(
+                            '/blog',
+                            [
+                                Route::get('/post')->name('api-v2-blog-post/index'),
+                                Route::get('/post/{id}')->name('api-v2-blog-post/view'),
+                            ]
+                        ),
+                        Route::get('/note')->name('api-v2-note/index'),
+                        Route::get('/note/{id}')->name('api-v2-note/view'),
+                    ]
+                )
+            ],
+            $container
+        );
+
+        $items = $apiGroup->getItems();
+
+        $this->assertAllRoutesAndGroupsHaveContainer($items);
+    }
+
+    public function testContainerAutoInjectionByGroup(): void
+    {
+        $container = $this->getContainer();
+
+        $apiGroup = new Group(
+            '/api', function (Group $group) {
+            $group->addRoute(Route::get('/info')->name('api-info'));
+            $group->addGroup(
+                new Group(
+                    '/v1', function (Group $group) {
+                    $group->addRoute(Route::get('/user')->name('api-v1-user/index'));
+                    $group->addRoute(Route::get('/user/{id}')->name('api-v1-user/view'));
+                    $group->addGroup(
+                        new Group(
+                            '/news', function (Group $group) {
+                            $group->addRoute(Route::get('/post')->name('api-v1-news-post/index'));
+                            $group->addRoute(Route::get('/post/{id}')->name('api-v1-news-post/view'));
+                        }
+                        )
+                    );
+                    $group->addGroup(
+                        new Group(
+                            '/blog', function (Group $group) {
+                            $group->addRoute(Route::get('/post')->name('api-v1-blog-post/index'));
+                            $group->addRoute(Route::get('/post/{id}')->name('api-v1-blog-post/view'));
+                        }
+                        )
+                    );
+                    $group->addRoute(Route::get('/note')->name('api-v1-note/index'));
+                    $group->addRoute(Route::get('/note/{id}')->name('api-v1-note/view'));
+                }
+                )
+            );
+            $group->addGroup(
+                new Group(
+                    '/v2', function (Group $group) {
+                    $group->addRoute(Route::get('/user')->name('api-v2-user/index'));
+                    $group->addRoute(Route::get('/user/{id}')->name('api-v2-user/view'));
+                    $group->addGroup(
+                        new Group(
+                            '/news', function (Group $group) {
+                            $group->addRoute(Route::get('/post')->name('api-v2-news-post/index'));
+                            $group->addRoute(Route::get('/post/{id}')->name('api-v2-news-post/view'));
+                        }
+                        )
+                    );
+                    $group->addGroup(
+                        new Group(
+                            '/blog', function (Group $group) {
+                            $group->addRoute(Route::get('/post')->name('api-v2-blog-post/index'));
+                            $group->addRoute(Route::get('/post/{id}')->name('api-v2-blog-post/view'));
+                        }
+                        )
+                    );
+                    $group->addRoute(Route::get('/note')->name('api-v2-note/index'));
+                    $group->addRoute(Route::get('/note/{id}')->name('api-v2-note/view'));
+                }
+                )
+            );
+        }, $container
+        );
+
+        $items = $apiGroup->getItems();
+
+        $this->assertAllRoutesAndGroupsHaveContainer($items);
+    }
+
+    public function testContainerAutoInjectionByRouterFactory(): void
+    {
+        $container = $this->getContainer();
+
+        $routes = [
+            Route::get('/info')->name('api-info'),
+            Group::create(
+                '/v2',
+                [
+                    Route::get('/user')->name('api-v2-user/index'),
+                    Route::get('/user/{id}')->name('api-v2-user/view'),
+                    Group::create(
+                        '/news',
+                        [
+                            Route::get('/post')->name('api-v2-news-post/index'),
+                            Route::get('/post/{id}')->name('api-v2-news-post/view'),
+                        ]
+                    ),
+                    Group::create(
+                        '/blog',
+                        [
+                            Route::get('/post')->name('api-v2-blog-post/index'),
+                            Route::get('/post/{id}')->name('api-v2-blog-post/view'),
+                        ]
+                    ),
+                    Route::get('/note')->name('api-v2-note/index'),
+                    Route::get('/note/{id}')->name('api-v2-note/view'),
+                ]
+            ),
+            Group::create(
+                '/v2',
+                [
+                    Route::get('/user')->name('api-v2-user/index'),
+                    Route::get('/user/{id}')->name('api-v2-user/view'),
+                    Group::create(
+                        '/news',
+                        [
+                            Route::get('/post')->name('api-v2-news-post/index'),
+                            Route::get('/post/{id}')->name('api-v2-news-post/view'),
+                        ]
+                    ),
+                    Group::create(
+                        '/blog',
+                        [
+                            Route::get('/post')->name('api-v2-blog-post/index'),
+                            Route::get('/post/{id}')->name('api-v2-blog-post/view'),
+                        ]
+                    ),
+                    Route::get('/note')->name('api-v2-note/index'),
+                    Route::get('/note/{id}')->name('api-v2-note/view'),
+                ]
+            )
+        ];
+
+        $factory = new RouterFactory($this->getEngineFactory(), $routes);
+        $router = $factory($container);
+        $items = $router->getItems();
+
+        $this->assertAllRoutesAndGroupsHaveContainer($items);
+    }
+
+    private function assertAllRoutesAndGroupsHaveContainer($items): void
+    {
+        $func = function ($item) use (&$func) {
+            $this->assertTrue($item->hasContainer());
+            if ($item instanceof Group) {
+                $items = $item->getItems();
+                array_walk($items, $func);
+            }
+        };
+        array_walk($items, $func);
+    }
+
+    private function getContainer(array $instances = []): ContainerInterface
+    {
+        return new Container($instances);
+    }
+
+    private function getEngineFactory(): callable
+    {
+        return new class() {
+            public function __invoke(): RouterInterface
+            {
+                return new class() extends Group implements RouterInterface {
+                    public function match(ServerRequestInterface $request): MatchingResult
+                    {
+                    }
+
+                    public function generate(string $name, array $parameters = []): string
+                    {
+                    }
+
+                    public function getUriPrefix(): string
+                    {
+                    }
+
+                    public function setUriPrefix(string $name): void
+                    {
+                    }
+                };
+            }
+        };
+    }
+}
